+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed the second call to `IGMarkets::Session#sign_in` looping indefinitely if the previously logged in session was not
   explicitly signed out
+- Fixed automatic re-sign-in when the client security token is invalid not working correctly
 
 ### 0.17 — July 27, 2016